move pref change registrar to AdBlockService
