+++ browser/brave_shields/ad_block_pref_service_factory.cc
@@ -48,8 +48,7 @@ AdBlockPrefServiceFactory::BuildServiceInstanceForBrowserContext(
   Profile* profile = Profile::FromBrowserContext(context);
 
   auto service = std::make_unique<AdBlockPrefService>(
-      g_brave_browser_process->ad_block_service(), profile->GetPrefs(),
-      g_browser_process->local_state(),
+      profile->GetPrefs(), g_browser_process->local_state(),
       g_browser_process->GetApplicationLocale());
 
   auto pref_proxy_config_tracker =
