+++ components/brave_shields/content/browser/ad_block_pref_service.cc
@@ -8,42 +8,18 @@
 #include <utility>
 
 #include "base/check.h"
-#include "base/functional/bind.h"
-#include "brave/components/brave_shields/content/browser/ad_block_service.h"
 #include "brave/components/brave_shields/core/browser/brave_shields_locale_utils.h"
-#include "brave/components/brave_shields/core/common/brave_shield_constants.h"
 #include "brave/components/brave_shields/core/common/pref_names.h"
 #include "components/prefs/pref_change_registrar.h"
 #include "components/prefs/pref_service.h"
 #include "components/proxy_config/pref_proxy_config_tracker.h"
 #include "components/user_prefs/user_prefs.h"
-#include "content/public/browser/browser_context.h"
 
 namespace brave_shields {
 
-namespace {
-
-std::string GetTagFromPrefName(const std::string& pref_name) {
-  if (pref_name == prefs::kFBEmbedControlType) {
-    return brave_shields::kFacebookEmbeds;
-  }
-  if (pref_name == prefs::kTwitterEmbedControlType) {
-    return brave_shields::kTwitterEmbeds;
-  }
-  if (pref_name == prefs::kLinkedInEmbedControlType) {
-    return brave_shields::kLinkedInEmbeds;
-  }
-  return "";
-}
-
-}  // namespace
-
-AdBlockPrefService::AdBlockPrefService(AdBlockService* ad_block_service,
-                                       PrefService* prefs,
+AdBlockPrefService::AdBlockPrefService(PrefService* prefs,
                                        PrefService* local_state,
-                                       const std::string& locale)
-    : ad_block_service_(ad_block_service), prefs_(prefs) {
-  pref_change_registrar_.reset(new PrefChangeRegistrar());
+                                       const std::string& locale) {
   if (!local_state->GetBoolean(
           prefs::kMigratedAdblockSocialMediaBlockingSettings)) {
     // Yes, we're reading these on the first profile launched, and migrating
@@ -59,24 +35,6 @@ AdBlockPrefService::AdBlockPrefService(AdBlockService* ad_block_service,
     local_state->SetBoolean(prefs::kMigratedAdblockSocialMediaBlockingSettings,
                             true);
   }
-  pref_change_registrar_->Init(prefs_);
-  pref_change_registrar_->Add(
-      prefs::kFBEmbedControlType,
-      base::BindRepeating(&AdBlockPrefService::OnPreferenceChanged,
-                          base::Unretained(this), prefs::kFBEmbedControlType));
-  pref_change_registrar_->Add(
-      prefs::kTwitterEmbedControlType,
-      base::BindRepeating(&AdBlockPrefService::OnPreferenceChanged,
-                          base::Unretained(this),
-                          prefs::kTwitterEmbedControlType));
-  pref_change_registrar_->Add(
-      prefs::kLinkedInEmbedControlType,
-      base::BindRepeating(&AdBlockPrefService::OnPreferenceChanged,
-                          base::Unretained(this),
-                          prefs::kLinkedInEmbedControlType));
-  OnPreferenceChanged(prefs::kFBEmbedControlType);
-  OnPreferenceChanged(prefs::kTwitterEmbedControlType);
-  OnPreferenceChanged(prefs::kLinkedInEmbedControlType);
 
   ManageAdBlockOnlyModeByLocale(local_state, locale);
 }
@@ -107,8 +65,6 @@ AdBlockPrefService::GetLatestProxyConfig(
 }
 
 void AdBlockPrefService::Shutdown() {
-  pref_change_registrar_.reset();
-
   // `pref_proxy_config_tracker_` has a reference to `proxy_config_service_`,
   // therefore detach `pref_proxy_config_tracker_` first, to prevent the
   // reference from dangling.
@@ -122,15 +78,6 @@ void AdBlockPrefService::Shutdown() {
     proxy_config_service_->RemoveObserver(this);
     proxy_config_service_.reset();
   }
-}
-
-void AdBlockPrefService::OnPreferenceChanged(const std::string& pref_name) {
-  std::string tag = GetTagFromPrefName(pref_name);
-  if (tag.length() == 0) {
-    return;
-  }
-  bool enabled = prefs_->GetBoolean(pref_name);
-  ad_block_service_->EnableTag(tag, enabled);
 }
 
 void AdBlockPrefService::OnProxyConfigChanged(
