+++ components/brave_shields/content/browser/ad_block_pref_service.h
@@ -15,19 +15,15 @@
 #include "net/proxy_resolution/proxy_config_service.h"
 #include "net/proxy_resolution/proxy_config_with_annotation.h"
 
-class PrefChangeRegistrar;
 class PrefService;
 class PrefProxyConfigTracker;
 
 namespace brave_shields {
 
-class AdBlockService;
-
 class AdBlockPrefService : public KeyedService,
                            public net::ProxyConfigService::Observer {
  public:
-  explicit AdBlockPrefService(AdBlockService* ad_block_service,
-                              PrefService* prefs,
+  explicit AdBlockPrefService(PrefService* prefs,
                               PrefService* local_state,
                               const std::string& locale);
   ~AdBlockPrefService() override;
@@ -41,18 +37,12 @@ class AdBlockPrefService : public KeyedService,
  private:
   void Shutdown() override;
 
-  void OnPreferenceChanged(const std::string& pref_name);
   void OnDeveloperModeChanged();
 
   // net::ProxyConfigService::Observer:
   void OnProxyConfigChanged(
       const net::ProxyConfigWithAnnotation& config,
       net::ProxyConfigService::ConfigAvailability availability) override;
-
-  raw_ptr<AdBlockService> ad_block_service_ = nullptr;  // not owned
-  raw_ptr<PrefService> prefs_ = nullptr;                // not owned
-  std::unique_ptr<PrefChangeRegistrar, content::BrowserThread::DeleteOnUIThread>
-      pref_change_registrar_;
 
   std::unique_ptr<PrefProxyConfigTracker> pref_proxy_config_tracker_;
   std::unique_ptr<net::ProxyConfigService> proxy_config_service_;
